Disbursement and mandate upload issues fixed
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/Release-2.41.0@74603 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ pff-interface-bajaj/src/main/java/com/pennanttech/bajaj/process/ControlDumpRequestProcess.java
@@ -69,11 +69,11 @@ public class ControlDumpRequestProcess extends DatabaseDataEngine {
 		parmMap.addValue("monthStartDate", monthStartDate);
 		parmMap.addValue("monthEndDate", monthEndDate);
 
-		jdbcTemplate.query(sql.toString(), parmMap, new ResultSetExtractor<Integer>() {
+		jdbcTemplate.query(sql.toString(), parmMap, new ResultSetExtractor<Long>() {
 			MapSqlParameterSource	map	= null;
 
 			@Override
-			public Integer extractData(ResultSet rs) throws SQLException, DataAccessException {
+			public Long extractData(ResultSet rs) throws SQLException, DataAccessException {
 				while (rs.next()) {
 					executionStatus.setRemarks("processing the record " + totalRecords++);
 					processedCount++;
